refactor: CuratorsPicksEmerging section
